phase-3/backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.34.0
sqlmodel==0.0.31
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.18.0
pydantic==2.12.0
pydantic-settings==2.12.0
sqlalchemy==2.0.45
uuid==1.30
resend==2.3.0
python-dateutil==2.9.0
google-generativeai==0.4.1